Extract branch logic
