+++ public/app/features/provisioning/components/Shared/ResourceEditFormSharedFields.tsx
@@ -10,6 +10,7 @@ import { WorkflowOption } from 'app/features/provisioning/types';
 import { validateBranchName } from 'app/features/provisioning/utils/git';
 import { isGitProvider } from 'app/features/provisioning/utils/repositoryTypes';
 
+import { useBranchDropdownOptions } from '../../hooks/useBranchDropdownOptions';
 import { useLastBranch } from '../../hooks/useLastBranch';
 import { usePRBranch } from '../../hooks/usePRBranch';
 import { generateNewBranchName } from '../utils/newBranchName';
@@ -46,66 +47,12 @@ export const ResourceEditFormSharedFields = memo<DashboardEditFormSharedFieldsPr
     const prBranch = usePRBranch();
     const lastBranch = getLastBranch(repository?.name);
 
-    const branchOptions = useMemo(() => {
-      const options: Array<{ label: string; value: string; description?: string }> = [];
-      const addedBranches = new Set<string>();
-
-      const configuredBranch = repository?.branch;
-      const configuredPrefix = t(
-        'provisioned-resource-form.save-or-delete-resource-shared-fields.suffix-configured-branch',
-        'Configured branch'
-      );
-      const prPrefix = t(
-        'provisioned-resource-form.save-or-delete-resource-shared-fields.suffix-pr-branch',
-        'Pull request branch'
-      );
-      const lastUsedPrefix = t(
-        'provisioned-resource-form.save-or-delete-resource-shared-fields.suffix-last-used',
-        'Last branch'
-      );
-
-      // 1. Show the configured branch first in the list
-      if (configuredBranch) {
-        options.push({
-          label: `${configuredBranch}`,
-          value: configuredBranch,
-          description: configuredPrefix,
-        });
-        addedBranches.add(configuredBranch);
-      }
-
-      // 2. Show the PR branch (from ref query param)
-      if (prBranch && !addedBranches.has(prBranch)) {
-        options.push({
-          label: prBranch,
-          value: prBranch,
-          description: prPrefix,
-        });
-        addedBranches.add(prBranch);
-      }
-
-      // 3. Show the last used branch
-      if (lastBranch && !addedBranches.has(lastBranch)) {
-        options.push({
-          label: lastBranch,
-          value: lastBranch,
-          description: lastUsedPrefix,
-        });
-        addedBranches.add(lastBranch);
-      }
-
-      // 4. Add other branches from the API
-      if (branchData?.items) {
-        for (const ref of branchData.items) {
-          if (!addedBranches.has(ref.name)) {
-            options.push({ label: ref.name, value: ref.name });
-            addedBranches.add(ref.name);
-          }
-        }
-      }
-
-      return options;
-    }, [branchData?.items, repository?.branch, prBranch, lastBranch]);
+    const branchOptions = useBranchDropdownOptions({
+      repository,
+      prBranch,
+      lastBranch,
+      branchData,
+    });
 
     const newBranchDefaultName = useMemo(() => generateNewBranchName(resourceType), [resourceType]);
